Refactored stretch range and legacy processing to separate files
